+++ pkg/proxy/iptables/proxier_test.go
@@ -703,7 +703,7 @@ func TestExternalIPsReject(t *testing.T) {
 				Name:       svcPortName.Port,
 				Port:       int32(svcPort),
 				Protocol:   api.ProtocolTCP,
-				TargetPort: intstr.FromInt(svcPort),
+				TargetPort: intstr.FromInt(int32(svcPort)),
 			}}
 		}),
 	)
@@ -946,7 +946,7 @@ func addTestPort(array []api.ServicePort, name string, protocol api.Protocol, po
 		Protocol:   protocol,
 		Port:       port,
 		NodePort:   nodeport,
-		TargetPort: intstr.FromInt(targetPort),
+		TargetPort: intstr.FromInt(int32(targetPort)),
 	}
 	return append(array, svcPort)
 }
